Debugging and confirming final project can work

+++ src/main/scala/cse250/pa0/DataProcessor.scala
@@ -27,5 +27,5 @@ object DataProcessor {
 
   def computeUniqueInverterManufacturers(dataset: Array[SolarInstallation]): Int = ???
 
-  def computeTotalExpectedKWHAnnualProduction(dataset: Array[SolarInstallation]): Int = ???
+  def computeTotalExpectedKWHAnnualProduction(dataset: Array[SolarInstallation]): Float = ???
 }
